feat: moved value Objects to an external folder
