+++ src/interfaces/entity.interface.ts
