+++ src/interfaces/valueObject.ts
