+++ src/valueObjects/index.ts
@@ -1,2 +1,3 @@
 export * from "./cpf";
 export * from "./email";
+export * from "./valorTotal";
